removing preprocess folder, and downloading 17lands data to dataloading utility file
